requirements.txt
asgiref==3.4.1
Django==3.2.9
django-tinymce==3.4.0
django-webpack-loader==0.7.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
